Fixed style errors
